kotlin: factor out db setup extension
